Migrando responsabilidad de infra a app

+++ src/main/java/org/uniremington/application/service/PerfilService.java
@@ -5,6 +5,7 @@ import jakarta.inject.Inject;
 import org.uniremington.application.service.interfaces.IPerfil;
 import org.uniremington.domain.model.Perfil;
 import org.uniremington.domain.repository.PerfilRepository;
+import org.uniremington.shared.exception.NotFoundException;
 
 import java.util.List;
 import java.util.Optional;
@@ -21,21 +22,61 @@ public class PerfilService implements IPerfil {
 
     @Override
     public Optional<Perfil> findById(Long id) {
-        return repository.findById(id);
+
+        Optional<Perfil> result = repository.findById(id);
+
+        if(result.isPresent()){
+            return result;
+        }
+
+        throw new NotFoundException("No se ha encontrado un perfil con el id:" + id);
+
     }
 
     @Override
     public List<Perfil> findAll() {
-        return repository.findAll();
+
+        List<Perfil> result = repository.findAll();
+
+        if(!result.isEmpty()){
+            return result;
+        }
+
+        throw new NotFoundException("No se han encontrado perfiles almacenados");
+
     }
 
     @Override
     public Perfil save(Perfil entity) {
-        return repository.save(entity);
+
+        if(entity.getId() == null){
+            throw new IllegalArgumentException("La cédula no puede ser nula al guardar un perfil");
+        }
+
+        Optional<Perfil> existente = repository.findById(entity.getId());
+
+        if (existente.isPresent()) {
+            return repository.save(entity, "update");
+        }
+
+        return repository.save(entity, "save");
+
     }
 
     @Override
     public void deleteById(Long id) {
-        repository.deleteById(id);
+
+        if(id == null){
+            throw new IllegalArgumentException("La cédula no puede ser nula al eliminar un perfil");
+        }
+
+        Optional<Perfil> existente = repository.findById(id);
+
+        if(existente.isPresent()){
+            repository.deleteById(id);
+        }
+
+        throw new NotFoundException("No se ha encontrado un perfil con el id:" + id);
+
     }
 }
